requirements.txt
numpy>=1.18.5
pandas>=0.23.4
umap-learn>=0.5.1
scikit-learn>=0.23.2
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
seaborn>=0.9.0
matplotlib>=2.2.3
scipy==1.4.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
